isl_to_text/requirements.txt
flask
flask-cors
tensorflow
pillow
numpy
opencv-python
gunicorn
fastapi
uvicorn
tensorflow
numpy
pillow
